requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
Pillow
opencv-python
python-multipart
gdown
lime